Rename CodeProber.jar and release via Github
This commit makes two long overdue and slightly breaking changes:

1) Renames "code-prober.jar" and "code-prober-dev.jar" to "CodeProber.jar"
2) Changes release management to be based on normal github releases,
   instead of the previous "VERSION" file and jar in git history.

The tool is called CodeProber, not Code-Prober, so the old jar name was
simply incorrect. Also, separating release and dev builds is
unnecessary, as that metadata is already present inside the build itself
(in cpr.properties).

Switching to github releases should stop the size of the repository from
growing too quickly.

+++ protocol/build.sh
@@ -7,15 +7,20 @@ mkdir build_tmp
 echo "Gathering sources.."
 find src -name "*.java" > sources.txt
 
-echo "Building.."
-javac @sources.txt -cp ../code-prober.jar -d build_tmp -source 8 -target 8
+if [ ! -f "../CodeProber.jar" ]; then
+  echo "Missing CodeProber.jar, it is needed to compile the protocol."
+  echo "Please run the top build.sh script first, or download the latest release and place it in the root of this repo"
+  exit 1
+fi
 
+echo "Building.."
+javac @sources.txt -cp ../CodeProber.jar -d build_tmp -source 8 -target 8
 
 java \
   -DJAVA_DST_DIR="../server/src/codeprober/protocol/data/" \
   -DJAVA_DST_PKG="codeprober.protocol.data" \
   -DTS_DST_FILE="../client/ts/src/protocol.ts" \
-  -cp build_tmp:../code-prober.jar protocolgen.GenAll
+  -cp build_tmp:../CodeProber.jar protocolgen.GenAll
 
 echo "Cleaning up.."
 rm sources.txt
